refactor(xod-client): move createPinWidgetsConfig into WidgetsGroup, rename it into PinWidgetsGroup, fix componentWillReceiveProps
