+++ packages/xod-client/src/editor/components/NodeInspector.jsx
@@ -1,4 +1,3 @@
-import * as R from 'ramda';
 import React from 'react';
 import PropTypes from 'prop-types';
 import * as XP from 'xod-project';
@@ -6,7 +5,7 @@ import * as XP from 'xod-project';
 import { WIDGET_TYPE } from '../constants';
 import { NODE_PROPERTY_KIND, NODE_PROPERTY_KEY } from '../../project/constants';
 
-import WidgetsGroup from './WidgetsGroup';
+import PinWidgetsGroup from './PinWidgetsGroup';
 import Widgets, { getNodeWidgetConfig } from './inspectorWidgets';
 
 import { RenderableNode } from '../../types';
@@ -14,54 +13,6 @@ import sanctuaryPropType from '../../utils/sanctuaryPropType';
 import { getUtmSiteUrl } from '../../utils/urls';
 
 import * as MESSAGES from '../messages';
-
-// :: RenderablePin -> String
-const getWidgetKey = R.converge((id, key) => `${id}_${key}`, [
-  R.prop('nodeId'),
-  R.prop('key'),
-]);
-
-const getPinWidgetProps = R.applySpec({
-  entityId: R.prop('nodeId'),
-  kind: R.always(NODE_PROPERTY_KIND.PIN),
-  key: getWidgetKey,
-  keyName: XP.getPinKey,
-  type: XP.getPinType,
-  label: XP.getPinLabel,
-  value: R.prop('value'),
-  direction: XP.getPinDirection,
-  isConnected: R.prop('isConnected'),
-  isInvalid: R.prop('isInvalid'),
-  deducedType: R.prop('deducedType'),
-  isBindable: XP.isPinBindable,
-  isLastVariadicGroup: R.prop('isLastVariadicGroup'),
-  specializations: R.prop('specializations'),
-});
-
-// :: RenderableNode -> { components: {...}, props: {...} }
-const createPinWidgetsConfig = R.compose(
-  R.reduce(
-    (acc, renderablePin) => {
-      const widgetProps = getPinWidgetProps(renderablePin);
-
-      const { component, props } = getNodeWidgetConfig(widgetProps.type);
-
-      const widget = Widgets.composeWidget(component, props);
-
-      return R.compose(
-        R.assocPath(['components', widgetProps.key], widget),
-        R.assocPath(['props', widgetProps.key], widgetProps)
-      )(acc);
-    },
-    { components: {}, props: {} }
-  ),
-  R.apply(R.concat),
-  R.map(R.sort(R.ascend(XP.getPinOrder))),
-  R.juxt([R.filter(XP.isInputPin), R.filter(XP.isOutputPin)]),
-  XP.normalizeEmptyPinLabels,
-  R.values,
-  R.prop('pins')
-);
 
 const NodeLabelWidget = Widgets.composeWidget(
   getNodeWidgetConfig(WIDGET_TYPE.LABEL).component,
@@ -112,11 +63,7 @@ const NodeInspector = ({ node, onPropUpdate, onNodeSpecializationChanged }) => {
 
       {DeadNodeMessage}
 
-      <WidgetsGroup
-        entity={node}
-        createWidgetsConfig={createPinWidgetsConfig}
-        onPropUpdate={onPropUpdate}
-      />
+      <PinWidgetsGroup node={node} onPropUpdate={onPropUpdate} />
 
       <NodeLabelWidget
         entityId={nodeId}
